test(codegen-new): reorganiza testes por ASSUNTO + cobre URL.canParse
Revisão: nomes de teste devem descrever O QUE testam, não a origem ("backfill").
Os dois `builtin_regression*.rs` (genéricos) são divididos em arquivos TEMÁTICOS:

- array_ops.rs        — splice/toSpliced/fill/copyWithin/push/unshift/concat
                        variádicos, toSorted(cmp), sort(cmp), join/slice/toString
                        defaults, Array.from(mapper)
- string_methods.rs   — normalize/isWellFormed/toWellFormed, métodos em variável
                        string, indexOf/lastIndexOf fromIndex
- object_static.rs    — Object.hasOwn/is/fromEntries
- error_cause.rs      — Error(message, { cause })
- json_parse.rs       — JSON.parse reviver (transform + delete)
- numeric_coercion.rs — boolean ToNumber, -0 literal
- url.rs              — URL.canParse static overload por aridade (NOVO — era o
                        único fix do dia ainda sem teste persistido)

Cobertura auditada por 2 agentes independentes: TODAS as 27 features dos PRs de
hoje têm um teste que GENUINAMENTE as exercita. Sem mudança de código de produção.

Testes: 774/774 unit verdes.
