+++ crates/rts-codegen-new/src/front/run/tests/mod.rs
@@ -72,8 +72,13 @@ mod globals;
 mod heap_field;
 mod inspect;
 mod instanceof_dyn;
-mod builtin_regression;
-mod builtin_regression2;
+mod array_ops;
+mod error_cause;
+mod json_parse;
+mod numeric_coercion;
+mod object_static;
+mod string_methods;
+mod url;
 mod logical;
 mod loops;
 mod math_tagged;
